Move the test build operation classes to a dedicated package.
