+++ subprojects/testing-base/src/test/groovy/org/gradle/api/internal/tasks/testing/operations/TestListenerBuildOperationAdapterTest.groovy
@@ -14,13 +14,14 @@
  * limitations under the License.
  */
 
-package org.gradle.api.tasks.testing
+package org.gradle.api.internal.tasks.testing.operations
 
 import org.gradle.api.internal.tasks.testing.TestCompleteEvent
 import org.gradle.api.internal.tasks.testing.TestDescriptorInternal
-import org.gradle.api.internal.tasks.testing.TestListenerBuildOperationAdapter
 import org.gradle.api.internal.tasks.testing.TestStartEvent
 import org.gradle.api.internal.tasks.testing.logging.SimpleTestOutputEvent
+import org.gradle.api.tasks.testing.TestOutputEvent
+import org.gradle.api.tasks.testing.TestResult
 import org.gradle.internal.operations.BuildOperationIdFactory
 import org.gradle.internal.progress.BuildOperationDescriptor
 import org.gradle.internal.progress.BuildOperationListener
