+++ subprojects/testing-jvm/src/integTest/groovy/org/gradle/testing/TestListenerBuildOperationAdapterIntegrationTest.groovy
@@ -16,7 +16,7 @@
 
 package org.gradle.testing
 
-import org.gradle.api.internal.tasks.testing.ExecuteTestBuildOperationType
+import org.gradle.api.internal.tasks.testing.operations.ExecuteTestBuildOperationType
 import org.gradle.integtests.fixtures.AbstractIntegrationSpec
 import org.gradle.integtests.fixtures.BuildOperationsFixture
 import org.gradle.integtests.fixtures.TestResources
